+++ backend/internal/services/recipestepcompletionconditions/config.go
@@ -1,4 +1,4 @@
-package recipestepingredients
+package recipestepcompletionconditions
 
 import (
 	"context"
