+++ backend/internal/services/recipestepcompletionconditions/config_test.go
@@ -1,4 +1,4 @@
-package recipestepingredients
+package recipestepcompletionconditions
 
 import (
 	"context"
