+++ backend/internal/services/recipestepcompletionconditions/doc.go
@@ -1,4 +1,4 @@
 /*
-Package recipestepingredients provides a series of HTTP handlers for managing recipe step ingredients in a compatible database.
+package recipestepcompletionconditions provides a series of HTTP handlers for managing recipe step ingredients in a compatible database.
 */
-package recipestepingredients
+package recipestepcompletionconditions
