+++ backend/internal/services/recipestepcompletionconditions/http_helpers_test.go
@@ -1,4 +1,4 @@
-package recipestepingredients
+package recipestepcompletionconditions
 
 import (
 	"context"
