+++ backend/internal/services/recipestepcompletionconditions/service.go
@@ -1,4 +1,4 @@
-package recipestepingredients
+package recipestepcompletionconditions
 
 import (
 	"fmt"
