+++ backend/internal/services/recipestepcompletionconditions/service_test.go
@@ -1,4 +1,4 @@
-package recipestepingredients
+package recipestepcompletionconditions
 
 import (
 	"errors"
